+++ src/otx/core/data/utils/utils.py
@@ -274,8 +274,8 @@ def adapt_tile_config(tile_config: TileConfig, dataset: Dataset, task: OTXTaskTy
         dataset (Dataset): Datumaro dataset including all subsets
         task (Task): task type of the model
     """
-    if (train_dataset := dataset.subsets().get("train")) is not None:
-        stat = compute_robust_dataset_statistics(train_dataset, task)
+    if (train_dataset := dataset.subsets().get("train") or dataset.subsets().get("TRAINING")) is not None:
+        stat = compute_robust_dataset_statistics(train_dataset, task=task)
         max_num_objects = round(stat["annotation"]["num_per_image"]["max"])
         avg_size = stat["annotation"]["size_of_shape"]["avg"]
         min_size = stat["annotation"]["size_of_shape"]["robust_min"]