+++ tests/unit/core/types/test_export.py
@@ -17,6 +17,7 @@ def test_wrap(fxt_label_info, task_type):
 
     multilabel = False
     hierarchical = False
+    output_raw_scores = True
     confidence_threshold = 0.0
     iou_threshold = 0.0
     return_soft_prediction = False
@@ -27,6 +28,7 @@ def test_wrap(fxt_label_info, task_type):
     params = params.wrap(
         multilabel=multilabel,
         hierarchical=hierarchical,
+        output_raw_scores=output_raw_scores,
         confidence_threshold=confidence_threshold,
         iou_threshold=iou_threshold,
         return_soft_prediction=return_soft_prediction,
@@ -44,6 +46,7 @@ def test_wrap(fxt_label_info, task_type):
     assert metadata[("model_info", "return_soft_prediction")] == str(return_soft_prediction)
     assert metadata[("model_info", "soft_threshold")] == str(soft_threshold)
     assert metadata[("model_info", "blur_strength")] == str(blur_strength)
+    assert metadata[("model_info", "output_raw_scores")] == str(output_raw_scores)
 
     # Tile config
     assert ("model_info", "tile_size") in metadata