+++ CHANGELOG.md
@@ -2,12 +2,37 @@
 
 All notable changes to this project will be documented in this file.
 
-## \[unreleased\]
+## \[2.3.0\]
 
 ### New features
 
 - Add YOLOv9 model for Object Detection
   (https://github.com/openvinotoolkit/training_extensions/pull/3917)
+- Add OV inference for keypoint detection
+  (https://github.com/openvinotoolkit/training_extensions/pull/3970)
+- Add tiling for semantic segmentation
+  (https://github.com/openvinotoolkit/training_extensions/pull/3954)
+
+### Enhancements
+
+- Upgrade OV, MAPI, and NNCF dependencies
+  (https://github.com/openvinotoolkit/training_extensions/pull/3967)
+- Instance Segmentation Model refactoring
+  (https://github.com/openvinotoolkit/training_extensions/pull/3865)
+- Bump torch and lightning to 2.4.0 versions
+  (https://github.com/openvinotoolkit/training_extensions/pull/3843)
+- Add mAP metric to evaluate multilabel classification
+  (https://github.com/openvinotoolkit/training_extensions/pull/3985)
+
+### Bug fixes
+
+- Fix a wrong HPO log
+  (https://github.com/openvinotoolkit/training_extensions/pull/3972)
+
+## \[2.2.0\]
+
+### New features
+
 - Add RT-DETR model for Object Detection
   (https://github.com/openvinotoolkit/training_extensions/pull/3741)
 - Add Multi-Label & H-label Classification with torchvision models
@@ -45,15 +70,31 @@ All notable changes to this project will be documented in this file.
   (<https://github.com/openvinotoolkit/training_extensions/pull/3769>)
 - Refactoring `ConvModule` by removing `conv_cfg`, `norm_cfg`, and `act_cfg`
   (<https://github.com/openvinotoolkit/training_extensions/pull/3783>, <https://github.com/openvinotoolkit/training_extensions/pull/3816>, <https://github.com/openvinotoolkit/training_extensions/pull/3809>)
+- Support ImageFromBytes
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3948>)
+- Enable model export
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3952>)
+- Move templates from OTX1.X to OTX2.X
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3951>)
+- Include Geti arrow dataset subset names
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3962>)
+- Include full image with anno in case there's no tile in tile dataset
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3964>)
+- Add type checker in converter for callable functions (optimizer, scheduler)
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3968>)
 
 ### Bug fixes
 
 - Fix Combined Dataloader & unlabeled warmup loss in Semi-SL
-  (https://github.com/openvinotoolkit/training_extensions/pull/3723)
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3723>)
 - Revert #3579 to fix issues with replacing coco_instance with a different format in some dataset
-  (https://github.com/openvinotoolkit/training_extensions/pull/3753)
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3753>)
 - Add num_devices in Engine for multi-gpu training
-  (https://github.com/openvinotoolkit/training_extensions/pull/3778)
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3778>)
+- Add missing tile recipes and various tile recipe changes
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3942>)
+- Change categories mapping logic
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3946>)
 
 ## \[v2.1.0\]
 
@@ -191,6 +232,8 @@ All notable changes to this project will be documented in this file.
   (<https://github.com/openvinotoolkit/training_extensions/pull/3684>)
 - Fix MaskRCNN SwinT NNCF Accuracy Drop
   (<https://github.com/openvinotoolkit/training_extensions/pull/3685>)
+- Fix MaskRCNN SwinT NNCF Accuracy Drop By Adding More PTQ Configs
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3929>)
 
 ### Known issues